Split Feedback extension in two
The Feedback extension had to many responsibilities and lost it's debug
 related functions to the newly created Debug extension.

The Debug extension is tasked with returning wrappers around the php
 var_dump, var_export and print_r functions who are used in the
 templates when EB is in debug context. These functions do not come
 natively in Twig.
